# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: IndiaConnect Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List / search services
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: category_id
          in: query
          schema:
            type: integer
        - name: seller_id
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated services
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceListResponse"
    post:
      operationId: createService
      tags: [services]
      summary: Create a service listing (seller only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not an approved seller
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /services/featured:
    get:
      operationId: listFeaturedServices
      tags: [services]
      summary: List featured services for homepage
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: Featured services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get a single service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateService
      tags: [services]
      summary: Update a service (owner seller only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service (owner seller only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Inquiries ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # ── Auth schemas ────────────────────────────────────────────────────
...
    Service:
      type: object
      properties:
        id:
          type: integer
        seller_id:
          type: integer
        seller_name:
          type: string
        seller_city:
          type: string
        category_id:
          type: integer
        category_name:
          type: string
        title:
          type: string
        description:
          type: string
        price_min:
          type: ["number", "null"]
        price_max:
          type: ["number", "null"]
        price_unit:
          type: string
          enum: [per_hour, per_day, per_project, fixed, negotiable]
        images:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, inactive]
        is_featured:
          type: boolean
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
      required:
        - id
        - seller_id
        - seller_name
        - seller_city
        - category_id
        - category_name
        - title
        - description
        - price_unit
        - images
        - status
        - is_featured
        - tags
        - created_at
    ServiceInput:
      type: object
      required: [category_id, title, description, price_unit]
      properties:
        category_id:
          type: integer
        title:
          type: string
          minLength: 5
        description:
          type: string
          minLength: 10
        price_min:
          type: ["number", "null"]
        price_max:
          type: ["number", "null"]
        price_unit:
          type: string
          enum: [per_hour, per_day, per_project, fixed, negotiable]
        images:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        is_featured:
          type: boolean
    ServiceUpdate:
      type: object
      properties:
        category_id:
          type: integer
        title:
          type: string
        description:
          type: string
        price_min:
          type: ["number", "null"]
        price_max:
          type: ["number", "null"]
        price_unit:
          type: string
          enum: [per_hour, per_day, per_project, fixed, negotiable]
        images:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, inactive]
        is_featured:
          type: boolean
    ServiceListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Service"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        total_pages:
          type: integer
      required:
        - items
        - total
        - page
        - limit
        - total_pages

    # ── Inquiry schemas ─────────────────────────────────────────────────
...